+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG
 
+## 0.1.4
+* Fixed a bug related to laravel-mix on Windows OS: 
+[Github Issue](https://github.com/JeffreyWay/laravel-mix/issues/1126)
+* clean up of npm dependencies
+
+## 0.1.3
+* Wording changes
+* Docomentation
+
+## 0.1.2
+* Fixed missing integration of the generated css file in index.html
+
+## 0.1.1
+* Fixed a problem with the .gitignore template file
+
 ## 0.1.0
 * Basic implementation of core functionality
     * Webpack